+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure `testuser` isn't reapplied when `ApplyCluster` is called again (e.g. during the upgrade tests)
+
 ## [0.12.0] - 2023-11-02
 
 ### Changed